+++ packages/cookie/src/index.ts
@@ -69,12 +69,14 @@ export type SerializeOptions = Partial<{
   expires: Date
 }>
 
-export function serialize(name: string, val: string, { encode = encodeURIComponent, domain, secure, httpOnly, expires, path, ...opt }: SerializeOptions) {
+export function serialize(name: string, val: string, opt: SerializeOptions = {}) {
+  if (!opt.encode) opt.encode = encodeURI
+
   if (!fieldContentRegExp.test(name)) {
     throw new TypeError('argument name is invalid')
   }
 
-  const value = encode(val)
+  const value = opt.encode(val)
 
   if (value && !fieldContentRegExp.test(value)) {
     throw new TypeError('argument val is invalid')
@@ -92,35 +94,35 @@ export function serialize(name: string, val: string, { encode = encodeURICompone
     str += '; Max-Age=' + Math.floor(maxAge)
   }
 
-  if (domain) {
-    if (!fieldContentRegExp.test(domain)) {
+  if (opt.domain) {
+    if (!fieldContentRegExp.test(opt.domain)) {
       throw new TypeError('option domain is invalid')
     }
 
-    str += '; Domain=' + domain
+    str += '; Domain=' + opt.domain
   }
 
-  if (path) {
-    if (!fieldContentRegExp.test(path)) {
+  if (opt.path) {
+    if (!fieldContentRegExp.test(opt.path)) {
       throw new TypeError('option path is invalid')
     }
 
-    str += '; Path=' + path
+    str += '; Path=' + opt.path
   }
 
-  if (expires) {
-    if (typeof expires.toUTCString !== 'function') {
+  if (opt.expires) {
+    if (typeof opt.expires.toUTCString !== 'function') {
       throw new TypeError('option expires is invalid')
     }
 
-    str += '; Expires=' + expires.toUTCString()
+    str += '; Expires=' + opt.expires.toUTCString()
   }
 
-  if (httpOnly) {
+  if (opt.httpOnly) {
     str += '; HttpOnly'
   }
 
-  if (secure) {
+  if (opt.secure) {
     str += '; Secure'
   }