+++ packages/cookie/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @tinyhttp/cookie
 
+## 0.1.13
+
+### Patch Changes
+
+- Fix obligatory options
+
 ## 0.1.12
 
 ### Patch Changes